Refactor PracticeSearchIndex#find
This commit is a no-op as far as functionality goes.

Stop calling the Blurrily::Map `map`, because it's a slightly overloaded
term. Call it `practices_haystack` so we can easily add more haystacks
(ie practitioners).

Rearrange some of the methods to make it easy to add new methods for
dealing with these soon-to-exist haystacks.