Limit results to a maximum of 10
If we get 10 excellent matches from the practices haystack and 10 awful
matches from the practitioners haystack, we should only be showing the
10 excellent matches. Otherwise there might be a better 11th match in
the practices set than any of the practitioners we end up showing.

To sort this out, get enough results from each haystack to fill up the
maximum, and then show the best X from all of them.

By default Blurrily will return 10 results, so this matches that for a
default.