Improve interface of ExecJSRenderer
Do the wrapping of the component with data attributes within the class
so that `#render` returns HTML that's all ready to go for use with
client side React.

Expose `#render_static_markup` for when you want to render a component
but don't intend to use it with client side React.

Tidy up the class with a few private methods.

Remove the `ServerRendering` namespace.

Add a custom error class for rendering errors from ExecJS.

Simplify the `react_component` helper to work with the new interface.