Add context to GP lookup page
Provide information about the service this is part of, booking a blood
test.

Explain why we need to know about the GP practice.

Make it clearer that the hint is suggesting what to search for, not what
you will find as a result of searching.

Provide a route out for people who aren't registered with a GP practice.