Cap results to a max of 20
Previously we were getting a maximum of 10 results for each of practice
name, practice address, and practitioner name, and then combining them
all into a single list and sorting.

That meant that the first 10 results would be of pretty high quality,
but they could quickly fall off after that.

Getting 20 result of each type and then returning only the 20 best
results should improve the consistency of result quality.

It would probably be better to use a quality threshold instead of an
arbitrary count.