Make result ordering more stable
Ruby's `Enumerable#sort` method is unstable (it uses qsort under the
hood). As a result, the order of results with the same score can change
depending on the total length of the list, due to the different order
and number of comparisons.

When we load a new page of results to show 40 instead of 20, the top 20
may not be in the same order as they were previously. That gives a
really weird visual jitter, and makes paging through results a pretty
awful experience.

When sorting include the original index for comparison so the results
stay stable.

This is a bit of hack, but it's good enough for now. It doesn't resolve
issues with a result's score changing as more results get included, nor
with GP names being added to existing results, changing the amount of
space they take up on screen.